discovery: move remotePubFromChanInfo to gossiper, remove utils
